+++ components/internal/flow/nodes/predictornode.tsx
@@ -1,5 +1,5 @@
 import {Card as BlueprintCard, Elevation} from "@blueprintjs/core"
-import {Tooltip as AntdTooltip} from "antd"
+import {Tooltip as AntdTooltip, Modal} from "antd"
 import {Text as EvergreenText, InfoSignIcon, Popover, Position, Tab, Tablist, Tooltip} from "evergreen-ui"
 import {NextRouter, useRouter} from "next/router"
 import {useSession} from "next-auth/react"
@@ -115,6 +115,10 @@ const PredictorNodeComponent: FC<NodeProps<PredictorNodeData>> = (props) => {
 
     // Since predictors change
     const [taggedData, setTaggedData] = useState<DataTag>(null)
+
+    // Allows the trash icon to change color when hovered over
+    const [trashHover, setTrashHover] = useState(false)
+    const trashColor = trashHover ? "red" : null
 
     //Set the dropdown defaults here since the dropdown is created here
     const DEFAULT_CLASSIFIER_METRIC = Array.from(metrics.classifier.keys())[0]
@@ -936,14 +940,45 @@ const PredictorNodeComponent: FC<NodeProps<PredictorNodeData>> = (props) => {
                         id={`${flowPrefix}-delete-button${idExtension}`}
                         type="button"
                         className="hover:text-red-700 text-xs"
-                        onClick={() => {
-                            DeleteNode(NodeID)
+                        onClick={(event) => {
+                            event.preventDefault()
+                            Modal.confirm({
+                                title: (
+                                    <span id={`delete-confirm-${flowPrefix}-title${idExtension}`}>
+                                        Delete this Predictor node?
+                                    </span>
+                                ),
+                                content: (
+                                    <span id={`delete-confirm-${flowPrefix}-message${idExtension}`}>
+                                        The node will be removed permanently{" "}
+                                        <b id={`bold-tag-${flowPrefix}-${idExtension}`}>
+                                            along with any associated downstream nodes.
+                                        </b>{" "}
+                                        This cannot be undone.
+                                    </span>
+                                ),
+                                centered: true,
+                                closable: true,
+                                okButtonProps: {
+                                    id: `delete-confirm-${flowPrefix}-ok-button${idExtension}`,
+                                },
+                                okText: "Delete",
+                                onOk: async () => {
+                                    DeleteNode(NodeID)
+                                },
+                                cancelText: "Keep",
+                                cancelButtonProps: {
+                                    id: `delete-confirm-${flowPrefix}-cancel-button${idExtension}`,
+                                },
+                            })
                         }}
                     >
                         <AiFillDelete
                             id={`${flowPrefix}-delete-button-fill${idExtension}`}
                             size="15"
-                            color="red"
+                            color={trashColor}
+                            onMouseEnter={() => setTrashHover(true)}
+                            onMouseLeave={() => setTrashHover(false)}
                         />
                     </button>
                 </div>
